feat(login): enhanced step switcher and fixed minor layout details
