+++ src/components/LoginSteps/index.ts
@@ -1,0 +1,3 @@
+import LoginStep1 from "./LoginStep1";
+import LoginStep2 from "./LoginStep2";
+export { LoginStep1, LoginStep2 };
