+++ src/components/index.ts
@@ -1,0 +1,2 @@
+import { LoginStep1, LoginStep2 } from "./LoginSteps";
+export { LoginStep1, LoginStep2 };
